+++ block_test.go
@@ -175,8 +175,7 @@ func TestBlockSize(t *testing.T) {
 			testutil.Ok(t, blockInit.Close())
 		}()
 		expSizeInit = blockInit.Size()
-		actSizeInit, err := testutil.DirSize(blockInit.Dir())
-		testutil.Ok(t, err)
+		actSizeInit := testutil.DirSize(t, blockInit.Dir())
 		testutil.Equals(t, expSizeInit, actSizeInit)
 	}
 
@@ -185,7 +184,7 @@ func TestBlockSize(t *testing.T) {
 		testutil.Ok(t, blockInit.Delete(1, 10, labels.NewMustRegexpMatcher("", ".*")))
 		expAfterDelete := blockInit.Size()
 		testutil.Assert(t, expAfterDelete > expSizeInit, "after a delete the block size should be bigger as the tombstone file should grow %v > %v", expAfterDelete, expSizeInit)
-		actAfterDelete, err := testutil.DirSize(blockDirInit)
+		actAfterDelete := testutil.DirSize(t, blockDirInit)
 		testutil.Ok(t, err)
 		testutil.Equals(t, expAfterDelete, actAfterDelete, "after a delete reported block size doesn't match actual disk size")
 
@@ -199,8 +198,7 @@ func TestBlockSize(t *testing.T) {
 			testutil.Ok(t, blockAfterCompact.Close())
 		}()
 		expAfterCompact := blockAfterCompact.Size()
-		actAfterCompact, err := testutil.DirSize(blockAfterCompact.Dir())
-		testutil.Ok(t, err)
+		actAfterCompact := testutil.DirSize(t, blockAfterCompact.Dir())
 		testutil.Assert(t, actAfterDelete > actAfterCompact, "after a delete and compaction the block size should be smaller %v,%v", actAfterDelete, actAfterCompact)
 		testutil.Equals(t, expAfterCompact, actAfterCompact, "after a delete and compaction reported block size doesn't match actual disk size")
 	}