+++ wal/wal_test.go
@@ -408,10 +408,8 @@ func TestCompression(t *testing.T) {
 		testutil.Ok(t, os.RemoveAll(dirUnCompressed))
 	}()
 
-	uncompressedSize, err := testutil.DirSize(dirUnCompressed)
-	testutil.Ok(t, err)
-	compressedSize, err := testutil.DirSize(dirCompressed)
-	testutil.Ok(t, err)
+	uncompressedSize := testutil.DirSize(t, dirUnCompressed)
+	compressedSize := testutil.DirSize(t, dirCompressed)
 
 	testutil.Assert(t, float64(uncompressedSize)*0.75 > float64(compressedSize), "Compressing zeroes should save at least 25%% space - uncompressedSize: %d, compressedSize: %d", uncompressedSize, compressedSize)
 }